refactor: retrofitのserviceを取得するメソッドをGitHubServiceに移動
