--- app/src/main/java/xyz/goodistory/yumemiassignment/Common.kt
@@ -1,18 +1,0 @@
-package xyz.goodistory.yumemiassignment
-
-import android.content.Context
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
-
-class Common {
-    companion object {
-        fun getGitHubService(context: Context): GitHubService {
-            val retrofit = Retrofit.Builder()
-                .baseUrl(context.getString(R.string.api_endpoint))
-                .addConverterFactory(GsonConverterFactory.create())
-                .build()
-
-            return retrofit.create(GitHubService::class.java)
-        }
-    }
-}
