+++ app/src/main/java/xyz/goodistory/yumemiassignment/GitHubService.kt
@@ -1,6 +1,9 @@
 package xyz.goodistory.yumemiassignment
 
+import android.content.Context
 import retrofit2.Call
+import retrofit2.Retrofit
+import retrofit2.converter.gson.GsonConverterFactory
 import retrofit2.http.GET
 import retrofit2.http.Path
 
@@ -11,7 +14,16 @@ interface GitHubService {
     @GET("users/{login}")
     fun getUser(@Path("login") login: String): Call<User>
 
+    companion object {
+        fun getGitHubService(context: Context): GitHubService {
+            val retrofit = Retrofit.Builder()
+                .baseUrl(context.getString(R.string.api_endpoint))
+                .addConverterFactory(GsonConverterFactory.create())
+                .build()
 
+            return retrofit.create(GitHubService::class.java)
+        }
+    }
 
     data class User(
         val login: String,
